deduplicate entry-form submission code
