+++ fava/templates/_overlays.html
@@ -26,7 +26,7 @@
 <div id="transaction" class="transaction-overlay overlay-wrapper">
   <div class="overlay">
     <button type=button class="link close-overlay">({{ _('close') }})</button>
-    <form id="transaction-form" action="{{ url_for('json_api.add_entries') }}">
+    <form id="transaction-form">
       <h3>{{ _('New transaction') }}:</h3>
       {{ forms.Transaction() }}
       <div class="fieldset">
