+++ fava/templates/extract.html
@@ -37,7 +37,7 @@
   </div>
   {% endfor %}
   <div class="fieldset ingest-footer">
-    <button type="button" data-event="extract-submit" data-url="{{ url_for('json_api.add_entries') }}">{{ _('Import transactions') }}</button>
+    <button type="button" data-event="extract-submit">{{ _('Import entries') }}</button>
   </div>
 </form>
 {% endblock %}
